Allow users to eager_load_all
I want to allow eager loading since it allows the app to use process to load
files one and then copy the memory with COW so it uses less memory. The html
view requires tubby which is an optional dependency so I don't want to load it
unless you specifically require it.